feat: テナントおよびワークフロー関連のAPIエンドポイントの改善と依存関係のチェックスクリプトの追加
- テナント削除APIのルーティングを整形
- JWKS検証器の初期化時にエラーハンドリングを追加
- ページネーション結果の全件数をint64に変更
- 統合テストでのメンバーリポジトリの期待値を整形
- ワークフロー関連のレスポンスJSON変換時にエラーハンドリングを追加
- 新しいティア依存方向検証スクリプトを追加し、Rust、Go、TypeScriptの依存関係をチェック
- Dart/Flutterの依存関係チェックスクリプトをripgrepからPOSIX互換のfindに変更

+++ regions/business/accounting/server/rust/domain-master/src/adapter/grpc/domain_master_grpc.rs
@@ -382,7 +382,7 @@ fn parse_uuid(raw: &str, field_name: &str) -> Result<Uuid, Status> {
 
 fn pagination_result(total_count: usize, page: i32, page_size: i32) -> PaginationResult {
     PaginationResult {
-        total_count: total_count as i32,
+        total_count: total_count as i64,
         page,
         page_size,
         has_next: (page as usize * page_size as usize) < total_count,
