+++ regions/system/server/rust/auth/src/adapter/grpc/auth_grpc.rs
@@ -113,7 +113,7 @@ impl AuthGrpcService {
                 Ok(ListUsersResponse {
                     users: proto_users,
                     pagination: Some(PaginationResult {
-                        total_count: result.pagination.total_count as i32,
+                        total_count: result.pagination.total_count as i64,
                         page: result.pagination.page,
                         page_size: result.pagination.page_size,
                         has_next: result.pagination.has_next,
